requirements.txt
fastapi
uvicorn
python-multipart
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
pydantic
pypdf
